+++ packages/coding-agent/src/tools/gh-cache-invalidation.ts
@@ -18,6 +18,7 @@
  * dwarfs the cost of one cache miss.
  */
 import { invalidateAllForNumber, invalidateAllForRepo } from "./github-cache";
+import { tokenizeShellSegments } from "./shell-tokenize";
 
 const PR_URL_PATTERN = /^https:\/\/github\.com\/([^/\s]+\/[^/\s]+)\/pull\/(\d+)(?:[/?#].*)?$/i;
 const ISSUE_URL_PATTERN = /^https:\/\/github\.com\/([^/\s]+\/[^/\s]+)\/issues\/(\d+)(?:[/?#].*)?$/i;
@@ -155,94 +156,13 @@ function detectGhMutation(tokens: readonly string[]): { number?: number; repo?: 
 }
 
 /**
- * Conservative tokenizer that splits a bash command into individual word
- * tokens. Handles single/double-quoted strings, backslash escapes, and
- * standard operators (`;`, `&&`, `||`, `|`, `&`, newlines) as token
- * boundaries that emit a sentinel `";"` so the caller treats the segments
- * as independent command sequences. We do not attempt full POSIX shell
- * parsing — heredocs, command substitution, and arithmetic expansion are
- * out of scope; the detector simply falls through when it cannot find a
- * clean `gh issue|pr <subcmd>` triple.
- */
-function tokenize(command: string): string[][] {
-	const segments: string[][] = [];
-	let current: string[] = [];
-	let buffer = "";
-	let inSingle = false;
-	let inDouble = false;
-	const pushBuffer = () => {
-		if (buffer.length > 0) {
-			current.push(buffer);
-			buffer = "";
-		}
-	};
-	const pushSegment = () => {
-		pushBuffer();
-		if (current.length > 0) segments.push(current);
-		current = [];
-	};
-	for (let i = 0; i < command.length; i++) {
-		const ch = command[i];
-		if (inSingle) {
-			if (ch === "'") {
-				inSingle = false;
-				continue;
-			}
-			buffer += ch;
-			continue;
-		}
-		if (inDouble) {
-			if (ch === "\\" && i + 1 < command.length) {
-				const next = command[i + 1];
-				if (next === '"' || next === "\\" || next === "$" || next === "`") {
-					buffer += next;
-					i++;
-					continue;
-				}
-			}
-			if (ch === '"') {
-				inDouble = false;
-				continue;
-			}
-			buffer += ch;
-			continue;
-		}
-		if (ch === "'") {
-			inSingle = true;
-			continue;
-		}
-		if (ch === '"') {
-			inDouble = true;
-			continue;
-		}
-		if (ch === "\\" && i + 1 < command.length) {
-			buffer += command[i + 1];
-			i++;
-			continue;
-		}
-		if (ch === " " || ch === "\t") {
-			pushBuffer();
-			continue;
-		}
-		if (ch === "\n" || ch === ";" || ch === "&" || ch === "|" || ch === "(" || ch === ")") {
-			pushSegment();
-			// `&&`, `||` already collapsed by the segment break above.
-			continue;
-		}
-		buffer += ch;
-	}
-	pushSegment();
-	return segments;
-}
-
-/**
  * Drop `github-cache` rows for any `gh issue|pr <mutating-subcmd>` call
  * embedded in `command`. Safe to invoke unconditionally; no-op when the
  * command does not touch GitHub state.
  */
 export function invalidateGithubCacheForBashCommand(command: string): void {
 	if (!command?.includes("gh")) return;
-	const segments = tokenize(command);
+	const segments = tokenizeShellSegments(command);
 	for (const segment of segments) {
 		const hit = detectGhMutation(segment);
 		if (!hit) continue;
